% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_plot.R
\name{boot_plot}
\alias{boot_plot}
\title{Plot Bootstrap Distributions}
\usage{
boot_plot(dens, title = "Distribution")
}
\arguments{
\item{dens}{An object of class 'exboot', or 'regboot'}

\item{title}{A plot title}
}
\value{
A ggplot object showing the density estimates of a bootstrap sample
statistic.
}
\description{
Generates a plot of the density estimates of a bootstrap
distribution.
}
\seealso{
\code{\link{exact_bootstrap}}
}
\keyword{internal}
