% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootsummer.R
\name{bootsummer}
\alias{bootsummer}
\title{Bootstrap Summary Statistics}
\usage{
bootsummer(dens, lb = 0.025, ub = 0.975)
}
\arguments{
\item{dens}{A list with components x (the coordinates of the points where the
density is estimated)
and y (the estimated density values), which gives the density estimate of the
bootstrap sample statistics.}

\item{lb}{Lower bound for the highest density interval. Default is 0.025.}

\item{ub}{Upper bound for the highest density interval. Default is 0.975.}
}
\value{
A list with components mode (the mode of the density estimate),
mean (the mean of the density estimate), sd (the standard deviation of the
density estimate), and HDI (the highest density interval of the density estimate).
}
\description{
This function computes the mode, mean, standard deviation, and highest density
interval of the density estimate of a bootstrap sample.
}
\examples{
set.seed(123)
data <- rnorm(5)
boot_result <- reg_bootstrap(data)
bootsummer(boot_result)
}
