% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecase_bootstrap.R
\name{ecase_bootstrap}
\alias{ecase_bootstrap}
\title{Performs an Exact Case Bootstrap with Summary Statistics}
\usage{
ecase_bootstrap(
  data,
  check_size = TRUE,
  anon = function(x) (mean(x)),
  lb = 0.025,
  ub = 0.975,
  density_args
)
}
\arguments{
\item{data}{A numeric vector of data points.}

\item{check_size}{A logical value indicating whether to check if the size of the
dataset is less than 10. Default is TRUE.}

\item{anon}{A function to compute the sample statistic for each bootstrap
sample. Default is the mean function.}

\item{lb}{Lower bound for the confidence interval. Default is 0.025.}

\item{ub}{Upper bound for the confidence interval. Default is 0.975.}

\item{density_args}{Pass additional arguments to stats::density}
}
\value{
A list with two elements: dens (a density estimate of the bootstrap
sample statistics) and stats (a list of summary statistics including mode,
median, mean, standard deviation, lower confidence interval (lCI), and upper
confidence interval (uCI)).
}
\description{
Performs an exact case bootstrap method on small datasets and returns
a set of summary statistics. For a given sample size N, it generates all
(2N - 1) choose N combinations with replacement. This is sometimes referred to
as exact case bootstrap resampling.
}
\examples{
set.seed(123)
data <- rnorm(5)
result <- ecase_bootstrap(data)
summary(result)
plot(result)
}
\references{
Jekel, C. F., & Romero, V. J. (2020). Conservative Estimation of Tail Probabilities from Limited Sample Data. \doi{https://doi.org/10.2172/1605343}
}
\seealso{
\code{\link{plot.ecboot}}, \code{\link{summary.ecboot}}
}
