% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_bootstrap.R
\name{exact_bootstrap}
\alias{exact_bootstrap}
\title{Exact Bootstrap}
\usage{
exact_bootstrap(
  data,
  n_bootstraps = 10000,
  check_size = TRUE,
  anon = function(x) (mean(x))
)
}
\arguments{
\item{data}{A numeric vector of data points.}

\item{n_bootstraps}{An integer indicating the number of bootstrap samples to
generate. Default is 10000.}

\item{check_size}{A logical value indicating whether to check if the size of
the dataset is less than 10. Default is TRUE.}

\item{anon}{A function to compute the sample statistic for each bootstrap
sample. Default is the mean function.}
}
\value{
A list with components x (the n coordinates of the points where the
density is estimated) and y (the estimated density values) which gives the
density estimate of the bootstrap sample statistics.
}
\description{
This function performs an exact bootstrap for small datasets of less than 10
observations.
}
\examples{
set.seed(123)
data <- rnorm(5)
exact_bootstrap(data)
}
