% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{exactamente_app}
\alias{exactamente_app}
\title{exactamente_app: Interactive Shiny App for Exploring Bootstrap Methods}
\usage{
exactamente_app()
}
\value{
A shinyApp object representing the running Shiny application. This function
is primarily called for its side effects, which include the launching and running
of the Shiny app.
}
\description{
Launches a Shiny app to interactively explore and compare Exact and Regular
bootstrap methods and All. The app provides a user-friendly interface, allowing
users to modify various parameters associated with each bootstrap method,
visualize the results, and understand the differences between the methods.
}
\examples{
if(interactive()){
  exactamente_app()
}
}
