% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecase_bootstrap.R
\name{plot.ecboot}
\alias{plot.ecboot}
\title{Plot Method for 'ecboot' Class}
\usage{
\method{plot}{ecboot}(x, title = "Exact Case Bootstrap Distribution", ...)
}
\arguments{
\item{x}{An object of class 'ecboot', usually the output of the \code{\link{ecase_bootstrap}} function.}

\item{title}{A plot title. Default is "Exact Case Bootstrap Distribution".}

\item{...}{Additional parameters (currently ignored).}
}
\value{
A ggplot object showing the density estimates of the bootstrap sample statistic.
}
\description{
Creates a plot of the density estimates of the bootstrap sample statistics
returned from the \code{\link{ecase_bootstrap}} function.
}
\examples{
set.seed(123)
data <- rnorm(5)
result <- ecase_bootstrap(data)
plot(result)
}
\seealso{
\code{\link{ecase_bootstrap}}, \code{\link{summary.ecboot}}
}
