% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecase_bootstrap.R
\name{summary.ecboot}
\alias{summary.ecboot}
\title{Summary Method for 'ecboot' Class}
\usage{
\method{summary}{ecboot}(object, ...)
}
\arguments{
\item{object}{An object of class 'ecboot', usually the output of the \code{\link{ecase_bootstrap}} function.}

\item{...}{Additional parameters (currently ignored).}
}
\value{
A data.frame containing the summary statistics.
}
\description{
Creates a summary table of the summary statistics computed in the
\code{\link{ecase_bootstrap}} function.
}
\examples{
set.seed(123)
data <- rnorm(5)
result <- ecase_bootstrap(data)
summary(result)
}
\seealso{
\code{\link{ecase_bootstrap}}, \code{\link{plot.ecboot}}
}
