% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_bootstrap.R
\name{summary.regboot}
\alias{summary.regboot}
\title{Summary Method for 'regboot' Class}
\usage{
\method{summary}{regboot}(object, ...)
}
\arguments{
\item{object}{An object of class 'regboot', usually the output of the \code{\link{reg_bootstrap}} function.}

\item{...}{Additional parameters (currently ignored).}
}
\value{
A data.frame containing the summary statistics.
}
\description{
Creates a summary table of the summary statistics computed in the
\code{\link{reg_bootstrap}} function.
}
\examples{
set.seed(123)
data <- rnorm(5)
result <- reg_bootstrap(data)
summary(result)
}
\seealso{
\code{\link{reg_bootstrap}}, \code{\link{plot.regboot}}
}
