% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{xboot_summary}
\alias{xboot_summary}
\title{Generate a Summary Table for Bootstrap Objects}
\usage{
xboot_summary(object)
}
\arguments{
\item{object}{A bootstrap object of class 'extboot', 'ecboot' or 'regboot'.}
}
\value{
A data frame containing the summary statistics with the method as the
first column.
}
\description{
This function takes a bootstrap object and generates a summary table
containing the method used and the computed summary statistics.
}
\keyword{internal}
