% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_plot.R
\name{boot_plot}
\alias{boot_plot}
\title{Plot Bootstrap Distributions}
\usage{
boot_plot(dens, title = "Distribution")
}
\arguments{
\item{dens}{A list with components x (the coordinates
of the points where the density is estimated) and y (the estimated density
values), representing the density estimate of a bootstrap sample statistics.}

\item{title}{A plot title}
}
\value{
A ggplot object showing the density estimates of a bootstrap sample
statistic.
}
\description{
This function generates a plot comparing of the density estimates of a bootstrap
distribution.
}
\examples{
set.seed(123)
data <- rnorm(5)
exact_bootstrap_result <- exact_bootstrap(data)
boot_plot(exact_bootstrap_result)
}
