% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_plot.R
\name{boot_plot}
\alias{boot_plot}
\title{Plot Bootstrap Distributions}
\usage{
boot_plot(dens, title = "Distribution")
}
\arguments{
\item{dens}{An object of class \code{ecboot}, \code{exboot}, or \code{regboot}}

\item{title}{A plot title}
}
\value{
A ggplot object showing the density estimates of a bootstrap sample
statistic.
}
\description{
Generates a plot of the density estimates of a bootstrap
distribution.
}
\examples{
\dontrun{
#set.seed(123)
#data <- rnorm(5)
#result <- exact_bootstrap(data)
#boot_plot(result)
}
}
\seealso{
\code{\link{exact_bootstrap}}
}
\keyword{internal}
