% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactamente_app.R
\name{exactamente_app}
\alias{exactamente_app}
\title{Run the Exactamente Shiny App}
\usage{
exactamente_app()
}
\value{
Nothing; this function is called for its side effect of starting the
Shiny app.
}
\description{
This function starts a Shiny app that lets the user interactively run
bootstrap analysis using either exact or regular bootstrapping, or both.
}
\examples{
\dontrun{
exactamento_app()
}
}
