% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_bootstrap.R
\name{reg_bootstrap}
\alias{reg_bootstrap}
\title{Regular Bootstrap}
\usage{
reg_bootstrap(data, n_bootstraps = 10000, anon = function(x) (mean(x)))
}
\arguments{
\item{data}{A numeric vector of data points.}

\item{n_bootstraps}{An integer indicating the number of bootstrap samples to
generate. Default is 10000.}

\item{anon}{A function to compute the sample statistic for each bootstrap
sample. Default is the mean function.}
}
\value{
A list with components x (the n coordinates of the points where the
density is estimated) and y (the estimated density values) which gives the
density estimate of the bootstrap sample statistics.
}
\description{
This function performs a regular bootstrap to generate a distribution of sample
statistics.
}
\examples{
set.seed(123)
data <- rnorm(5)
reg_bootstrap(data)
}
